[Feat] 소켓 준비 및 메모리 누수 방지 Refactor

+++ front_android/lib/theme/components/progress_bar.dart
@@ -10,8 +10,8 @@ class ProgressBar extends StatelessWidget {
     this.flipHorizontally = false,
   });
 
-  final int currentProgress;
-  final int fullProgress;
+  final double currentProgress;
+  final double fullProgress;
   final Color valueColor;
   final Color backgroundColor;
   final bool flipHorizontally;
